services/exception-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
tortoise-orm==0.20.0
aerich==0.7.2
asyncpg==0.29.0
faker==22.6.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0